requirements.txt
Django==1.9.2
pinax-theme-bootstrap==7.3.0
django-user-accounts==1.3.1
metron==1.3.7
pinax-eventlog==1.1.1
dj-database-url==0.4.0
gunicorn==19.4.5
psycopg2==2.6.1
pinax-pages==0.4.2
pinax-boxes==2.1.2
whitenoise==2.0.6

# run off of master for now
git+https://github.com/pinax/symposion.git

# DjangoCon custom requirements
django-markitup==2.3.1
pinax-blog==4.3.0
Unipath==1.1
